// bentador/src/types/customer-types.ts
export interface Customer {
  id: number
  name: string
  email: string
  joined: string
  orders: number
  completed_orders: number
  spent: string
  lv_raw: number
  last_login_at: string
  last_login_raw: string
  status: string
  color: string
  is_active: number
  type: string
}

// interface Customer {
//   name: string
//   email: string
//   joined: string
//   orders: string
//   spent: string
//   status?: string
//   color?: string
// }